base code tests reviewed, ain't sure if they are all rigth 1.1 instanceOf issues solved, at least for equipping items check the final keyword, write better tests. 1.2 study the codes is needed 1.3 idem 2.1 attacks implemented: errors ("tripple dispatch")

+++ src/model/units/Alpaca.java
@@ -24,7 +24,7 @@ public class Alpaca extends AbstractUnit {
    * @param location
    *     current position of the unit
    */
-  public Alpaca(int hitPoints, final int movement, Location location,
+  public Alpaca(double hitPoints, final int movement, Location location,
       IEquipableItem... items) {
     super(hitPoints, movement, location, Integer.MAX_VALUE, items);
   }
@@ -38,5 +38,7 @@ public class Alpaca extends AbstractUnit {
   public void equipItem(IEquipableItem item) {
     // Method body intentionally left empty
   }
+
+
 
 }
